feat: 优化EEG分类模型性能并增强实验功能
- 更新依赖版本要求至最低兼容版本
- 修改EEGDataset.py返回数据类型为float
- Classifier_Trainer.py中增加kappa和f1指标计算
- 重构CBAM模块为CTA模块，优化注意力机制
- 添加超参数搜索实验脚本，支持多折验证和结果记录
- 改进测试脚本，支持多次运行取最佳结果
- 新增实验日志记录功能，保存详细训练过程

+++ Utils/EEGDataset.py
@@ -73,7 +73,7 @@ class getSSVEP2020Intra(Dataset):
       eeg_data = torch.from_numpy(eeg_data.swapaxes(0, 1))  # (8, 180, 1024) -> (180, 8, 1024)
       eeg_data = eeg_data.reshape(-1, 1, self.Nc, self.Nt)  #(试次，1，NC,NT)符合CNN的输入
       print(eeg_data.shape)
-      return eeg_data
+      return eeg_data.float()
 
    # get the single label data
    def get_DataLabel(self):
